requirements.txt
pandas==2.2.3
numpy==1.26.4
crewai==0.102.0
autogen==0.4.1
pyautogen==0.10.0
langgraph==1.0.1
langchain==0.3.21
langchain-openai==0.2.12
pydantic==2.10.6
streamlit==1.52.1
tavily-python==0.7.14
python-dotenv==1.2.1
chromadb==0.5.23
uvicorn==0.38.0
fastapi==0.115.9
pytest==8.4.2
langflow==1.4.2
amadeus==12.0.0
boto3==1.34.162
botocore==1.34.162
deepeval==2.6.7